Added the first Certification Project (Build a Form)
Added the first Certification Project required to get the Responsive Web Design Certificate from freeCodeCamp. The project was to build a form that was similar in funcionality to this one: https://survey-form.freecodecamp.rocks/
I created an Order Form so that it would be different from the sample form and wouldn't seem like I copied everything.